all css removed, pictureGenerator removed and integrated into picture
