--- src/components/Pictures/picture.css
@@ -1,20 +1,0 @@
-.pictures-list {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-wrap: wrap;
-  padding: 0;
-  margin: 1rem;
-}
-.limitPictureLabel {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-bottom: 1rem;
-}
-form span {
-  margin-right: 0.5rem;
-}
-#limitPicture {
-  width: 250px;
-}
