--- src/components/Pictures/pictureCard.css
@@ -1,16 +1,0 @@
-li {
-  list-style-type: none;
-  text-decoration: none;
-}
-.card {
-  flex-flow: column wrap;
-  flex: 1 1 500px;
-  margin: 15px;
-  box-shadow: 0 0 10px rgba(51, 51, 51, 0.25);
-  overflow: hidden;
-  border-radius: 5px;
-}
-ul li {
-  margin: 1rem 0;
-  font-weight: 600;
-}
